# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "conexionPos",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/v1/warehouses": {
      "get": {
        "description": "get all warehouses of the BD",
        "tags": [
          "warehouse"
        ],
        "responses": {
          "200": {
            "description": "get all warehouse successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      },
      "post": {
        "description": "insert a new warehouse in the BD",
        "tags": [
          "warehouse"
        ],
        "requestBody": {
          "description": "warehouses create body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/warehouseSchema"
              },
              "example": {
                "name": "almacen prueba",
                "code": "prueba wareh",
                "location": "risaralda",
                "description": "almacen de prueba api",
                "warehouseTypes": "warehouse billing",
                "BillingResolutionId": "idbillingsdkjs"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "warehouse add successfully"
          },
          "401": {
            "description": "you can't create warehouse"
          }
        }
      }
    },
    "/v1/warehouses/{id}": {
      "get": {
        "description": "get one warehouse of the BD",
        "tags": [
          "warehouse"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the warehouse",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "get all warehouse successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      },
      "put": {
        "description": "update one warehouse of the BD",
        "tags": [
          "warehouse"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the warehouse",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "warehouses updated body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/warehouseSchema"
              },
              "example": {
                "name": "almacen prueba mod",
                "code": "prueba wareh mod",
                "location": "risaralda mod",
                "description": "almacen de prueba apimod",
                "warehouseTypes": "warehouse billing",
                "BillingResolutionId": "idbillingsdkjs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "update warehouse successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }
      },
      "delete": {
        "description": "delete one warehouse of the BD",
        "tags": [
          "warehouse"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the warehouse",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "delete warehouse successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      }
    },
...
    "/v1/monetaryDenomination": {
      "get": {
        "description": "get all monetaryDenominations of the BD",
        "tags": [
          "monetaryDenominations"
        ],
        "responses": {
          "200": {
            "description": "get all monetaryDenominations successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      },
      "post": {
        "description": "insert a new monetaryDenominations in the BD",
        "tags": [
          "monetaryDenominations"
        ],
        "requestBody": {
          "description": "monetaryDenominationss create body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/monetaryDenominationsSchema"
              },
              "example": {
                "photoFile": "idphototprueba",
                "value": 2000,
                "monetaryDenominationTypes": "COIN"
            }
            }
          }
        },
        "responses": {
          "201": {
            "description": "monetaryDenominations add successfully"
          },
          "401": {
            "description": "you can't create monetaryDenominations"
          }
        }
      }
    },
    "/v1/monetaryDenomination/{id}": {
      "get": {
        "description": "get one monetaryDenominations of the BD",
        "tags": [
          "monetaryDenominations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the monetaryDenominations",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "get all monetaryDenominations successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      },
      "put": {
        "description": "update one monetaryDenominations of the BD",
        "tags": [
          "monetaryDenominations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the monetaryDenominations",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "monetaryDenominations updated body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/monetaryDenominationsSchema"
              },
              "example": {
                "photoFile": "idphototprueba mod",
                "value": 3000,
                "monetaryDenominationTypes": "TICKET"
            }
            }
          }
        },
        "responses": {
          "200": {
            "description": "update monetaryDenominations successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }
      },
      "delete": {
        "description": "delete one monetaryDenominations of the BD",
        "tags": [
          "monetaryDenominations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the monetaryDenominations",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "delete monetaryDenominations successfully"
          },
          "401": {
            "description": "don't have permissions for this action"
          }
        }

      }
    }
  },
...
}